feat: vfs and initrd
Implementation of a virtual and initrd filesystems

Virtual filesystem is an API to manage underlying filesystems
(e.g. ext2, ntfs, fat, etc).

Initrd is an initial ramdisk.
It's purpose is providing necessary files for kernel to boot.
It has it's own underlying filesystem (initrdfs).
The loading is performed by GRUB through multiboot modules.
Due to lack of hardware drivers it is currently mounted as the root filesystem