feat: userland, syscalls and refactoring
Implementation of userland and ELF applications loader.
Userland is an address space that is setup by, but
does not belong to the kernel.
System calls (aka syscalls) is a way of communication between userland and kernel.
A syscall is initiated by calling interrupt 0x80 with
number of a syscall in eax register and
function arguments put in registers ebx, ecx, edx and others accordingly.
Most of the syscalls won't be compatible with linux kernel.
The compatibility mode can be enabled by defining
CONFIG_SYS_LINUX_COMPAT macro in depthos/kconfig.h.