feat: task scheduler and syscall tracing
Implement round-robin task scheduler and optional tracing for syscalls.
Task represents a running program with it's state.
Scheduler is a kernel module which suspends flow
of the current program and schedules execution of another.
The scheduler suspends the current program when it receives
a tick produced by PIT.
The current implementation also allows manual rescheduling through 0x30 interrupt.